Address search implicitly does radius search
